IntelliRAG-System/requirements.txt
langchain
langchain-community
langchain-huggingface
chromadb
pypdf
streamlit
python-dotenv
groq
sentence-transformers